% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{add_item_properties_mst}
\alias{add_item_properties_mst}
\title{Add item properties to an dextermst project}
\usage{
add_item_properties_mst(db, item_properties)
}
\arguments{
\item{db}{dexterMST project database}

\item{item_properties}{data.frame with a column item_id and other columns containing the item properties}
}
\description{
Add item properties to an dextermst project
}
