% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{add_person_properties_mst}
\alias{add_person_properties_mst}
\title{Add person properties to a mst project}
\usage{
add_person_properties_mst(db, person_properties)
}
\arguments{
\item{db}{dextermst project database}

\item{person_properties}{data.frame with a column person_id and 
other columns containing the person properties}
}
\description{
Add person properties to a mst project
}
