% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{add_scoring_rules_mst}
\alias{add_scoring_rules_mst}
\title{add scoring rules to an mst project}
\usage{
add_scoring_rules_mst(db, rules)
}
\arguments{
\item{db}{a dextermst db connection}

\item{rules}{dataframe (item_id, response, item_score), 
listing all permissible responses to an item and their scores}
}
\description{
add scoring rules to an mst project
}
