% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{alter_scoring_rules_mst}
\alias{alter_scoring_rules_mst}
\title{alter scoring rules in an mst project}
\usage{
alter_scoring_rules_mst(db, rules)
}
\arguments{
\item{db}{a dextermst db connection}

\item{rules}{data.frame (item_id, response, item_score), see dexter}
}
\description{
It is only possible to change item_scores for existing items and responses 
through this function. Scoring rules can only be changed for items that are 
in the last module of a (mst) test.
}
