% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enorm.R
\name{coef.mst_enorm}
\alias{coef.mst_enorm}
\title{extract enorm mst item parameters}
\usage{
\method{coef}{mst_enorm}(object, ...)
}
\arguments{
\item{object}{an 'mst_enorm' parameters object, generated by the function \code{\link{fit_enorm_mst}}}

\item{...}{other parameters are currently ignored}
}
\value{
a data.frame with columns: item_id, item_score, beta, SE_b
}
\description{
extract enorm mst item parameters
}
