% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{create_mst_project}
\alias{create_mst_project}
\title{create a new (empty) mst project}
\usage{
create_mst_project(pth)
}
\arguments{
\item{pth}{path and filename to save project file}
}
\value{
handle to project database
}
\description{
create a new (empty) mst project
}
