% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\docType{package}
\name{dexterMST-package}
\alias{dexterMST}
\alias{dexterMST-package}
\title{DexterMST: CML calibration and data management for multi stage tests}
\description{
DexterMST is a generalization of the most important functionality in dexter to multi stage tests.
Function names are typically the same as in dexter with '_mst' added. CML calibration of real life
mst tests is tricky, especially if one considers the need to condition on the design in combination with
data selections and corrections of key errors. DexterMST aims to handle these things automatically
and protect the user from making mistakes by working from a local database which enforces some restrictions.
}
\details{
The main features are:

\itemize{
\item project databases providing a structure for storing data about persons, items, responses and booklets.
\item CML calibration of the extended nominal response model and interaction model.
}

To learn more about dexterMST, start with the vignette: `vignette('dexterMST',package='dexterMST')`
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dexter-psychometrics.github.io/dexter/}
  \item Report bugs at \url{https://github.com/dexter-psychometrics/dexter/issues}
}

}
\author{
\strong{Maintainer}: Timo Bechger \email{tmbechger@gmail.com}

Authors:
\itemize{
  \item Jesse Koops
  \item Ivailo Partchev
  \item Gunter Maris
}

Other contributors:
\itemize{
  \item Robert Zwitser [contributor]
}

}
