% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{fit_inter_mst}
\alias{fit_inter_mst}
\title{Fit the interaction model on a single multi-stage booklet}
\usage{
fit_inter_mst(db, test_id, booklet_id)
}
\arguments{
\item{db}{a db handle}

\item{test_id}{id of the test as defined in \code{\link{create_mst_test}}}

\item{booklet_id}{id of the booklet as defined in \code{\link{create_mst_test}}}
}
\description{
Fit the interaction model on a single multi-stage booklet
}
