% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{get_responses_mst}
\alias{get_responses_mst}
\title{Extract response data from a dexterMST database}
\usage{
get_responses_mst(
  db,
  predicate = NULL,
  columns = c("person_id", "test_id", "booklet_id", "item_id", "item_score")
)
}
\arguments{
\item{db}{a dexterMST project database connection}

\item{predicate}{an expression to select data on}

\item{columns}{the columns you wish to select, can include any column in the project}
}
\value{
a data.frame of responses
}
\description{
Extract response data from a dexterMST database
}
