% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF.R
\name{plot.DIF_stats_mst}
\alias{plot.DIF_stats_mst}
\title{plot method for DIF_mst}
\usage{
\method{plot}{DIF_stats_mst}(x, items = NULL, itemsX = items, itemsY = items, ...)
}
\arguments{
\item{x}{object produced by DIF_mst}

\item{items}{character vector of item id's for a subset of the plot. Useful if you have many items. 
If NULL all items are plotted.}

\item{itemsX}{character vector of item id's for the X axis}

\item{itemsY}{character vector of item id's for the Y axis}

\item{...}{further arguments to plot}
}
\description{
plot method for DIF_mst
}
