% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{plot.im_mst}
\alias{plot.im_mst}
\title{plots for the interaction model}
\usage{
\method{plot}{im_mst}(x, item_id = NULL, show.observed = TRUE, curtains = 10, zoom = FALSE, ...)
}
\arguments{
\item{x}{output of \code{\link{fit_inter_mst}}}

\item{item_id}{id of the item to plot}

\item{show.observed}{plot the observed mean item scores for each test score}

\item{curtains}{percentage of most extreme values to cover with curtains, 0 to omit curtains}

\item{zoom}{if TRUE, limits the plot area to the test score range allowed by the routing rules}

\item{...}{further arguments to plot}
}
\description{
plots for the interaction model
}
