% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_analysis.R
\name{profile_tables_mst}
\alias{profile_tables_mst}
\title{Profile analysis}
\usage{
profile_tables_mst(parms, domains, item_property)
}
\arguments{
\item{parms}{An object returned by \code{\link{fit_enorm_mst}}}

\item{domains}{data.frame with column item_id and a column whose name matches `item_property`}

\item{item_property}{the name of the item property used to define the domains.}
}
\description{
Expected and observed domain scores per booklet and test score
}
