% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMST.R
\name{sim_mst}
\alias{sim_mst}
\title{Simulate multistage testing data}
\usage{
sim_mst(pars, theta, test_design, routing_rules, routing = c("last", "all"))
}
\arguments{
\item{pars}{item parameters, can be either: 
a data.frame with columns item_id, item_score, beta or a dexter or dexterMST parameters object}

\item{theta}{vector of person abilities}

\item{test_design}{data.frame with columns item_id, module_id, item_position}

\item{routing_rules}{output pf \code{\link{mst_rules}}}

\item{routing}{'all' or 'last' routing}
}
\description{
Simulates data from an extended nominal response model according to an mst design
}
