% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ability}
\alias{ability_tables}
\alias{plausible_values}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dexter}{\code{\link[dexter]{ability}}, \code{\link[dexter]{plausible_values}}}
}}

